Append new resource to resource collection to minimize impact.
Append the freshly created bin resource to the resource collection. Do
not prepend it. Prepend will shift indices in resource collection for
all other resources. This fixes an issue, where Chef 11.4.4 would not
find a module.

Appending will minimize the the impact of resource collection
modification. At the same time it wont harm.